requirements.txt
Mako==1.2.4
PyYAML==6.0
invoke
